Add create_before_destroy to parameter group (#30)
aws_db_parameter_group (or AWS?) does not support removing previously
defined parameters.

I thought I'd give "taint" a spin and force the parameter group to be
recreated.  terraform tried to delete the parameter group but failed as
it's still in use.

This change allows terraform to migrate the RDS instance to the new
parameter group before destroying the old.