Missing `ca_cert_identifier` for sqlserver engine aws rds instance
The recent PR (https://github.com/terraform-aws-modules/terraform-aws-rds/pull/174) included changes for the `ca_cert_identifier` variable, but it does not have it on the `aws_db_instance.this_mssql` resources (sqlserver engine types). This may be important for some AWS users. Also, we've briefly tested that the change can be applied without destroying the resource, but requires `apply-immediately` to be set .