update to follow change in aws_db_instance id
aws_db_instance was updated by https://github.com/hashicorp/terraform-provider-aws/pull/31232
such that the `id` attribute now tracks the resource ID, and the
identifier is now returned in a new attribute named `identifier`.

The resource ID is already available in `instance_resource_id`. This
fix allows existing users to continue using this module with no
changes and preserves the original function of the `id` attribute
in this module.

Fixes #495.